Allow multiple cli arguments to be passed and fix null field values (#214)
* Allow multiple cli arguments to be passed

This is needed if a project wants to add its own transformer.  This adds a test of an empty tranformer. Also this allows for arguments to be strings or arrays

* Fix Null field values

Now check if a value is null by checking if the reference is 0.

* nits, require breaks

* Add backup require for assemblyscript

Now if it isn't loaded already, use the normal require path

* Fix window paths

Doesn't assume unix paths

Co-authored-by: jtenner <tenner.joshua@gmail.com>