Fix CI (#400)
* Use actions/setup-node and update actions/checkout

dcodeIO/setup-node-nvm uses a version of nvm-windows that breaks npm,
hence the CI failures. There are two solutions: one, change the version
of nvm-windows on dcodeIO/setup-node-nvm, or two, replace the action
with actions/setup-node. I chose the latter.

* Test Node.js versions 18 and 19 on CI

CI should now test the LTS and current Node.js versions, which are 18
and 19 respectively. Node.js v16 is still being tested it is the minimum
version supported by asc, and therefore future versions of as-pect
should not break on Node.js v16.