Upgrade to as23 (#397)
* as23 start

* Update snapshot tests, scripts, and snapshots (#398)

This commit modifies the Jest config to properly use the ts-jest ESM
preset and remove redundant options. The snapshots have been updated
accordingly, and the snapshot update script in package.json has been
changed to use the same Node options as the regular test script.

* Update the strict equals transform to use Objects (#399)

The signature for __aspectStrictEquals has been updated to use an Object
as the first parameter, instead of the class type. The beginning of the
method now resembles the following pseudocode:

    class Foo {
      // ...

      __aspectStrictEquals(rawRef: Object, ...): bool {
        if (!(rawRef instanceof Foo)) return false;
        const ref = rawRef as Foo;

        // Check ref's fields...
      }
    }

This change should now allow __aspectStrictEquals to play nicely with
AssemblyScript v0.23.x.

* Lots of changes, need to wait for assemblyscript to drop with a few changes

* update snapshots

* Fix CI (#400)

* Use actions/setup-node and update actions/checkout

dcodeIO/setup-node-nvm uses a version of nvm-windows that breaks npm,
hence the CI failures. There are two solutions: one, change the version
of nvm-windows on dcodeIO/setup-node-nvm, or two, replace the action
with actions/setup-node. I chose the latter.

* Test Node.js versions 18 and 19 on CI

CI should now test the LTS and current Node.js versions, which are 18
and 19 respectively. Node.js v16 is still being tested it is the minimum
version supported by asc, and therefore future versions of as-pect
should not break on Node.js v16.

* everything passes

Co-authored-by: CountBleck <Mr.YouKnowWhoIAm@protonmail.com>